Add a directory to test java Stream API
First class TestTask.java includes methods that create a random
list of integers in which there is a every number exept a single
one has a counterpart, and then looking for alone number (without
a counterpart) in the list. List creating and looking for an alone
number provided with stream usage.